+++ app/src/main/java/com/example/teumteum/ui/calendar/CalendarMode.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum
+package com.example.teumteum.ui.main
 
 enum class CalendarMode {
     WEEKLY, MONTHLY
