+++ app/src/main/java/com/example/teumteum/ui/feed/FeedFragment.kt
@@ -1,22 +1,22 @@
-package com.example.teumteum
+package com.example.teumteum.ui.feed
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.teumteum.databinding.FragmentMyhomeBinding
+import com.example.teumteum.databinding.FragmentFeedBinding
 
-class MyhomeFragment : Fragment() {
+class FeedFragment : Fragment() {
 
-    lateinit var binding: FragmentMyhomeBinding
+    lateinit var binding: FragmentFeedBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentMyhomeBinding.inflate(inflater, container, false)
+        binding = FragmentFeedBinding.inflate(inflater, container, false)
 
         return binding.root
     }
