+++ app/src/main/java/com/example/teumteum/ui/friend/FriendFragment.kt
@@ -1,22 +1,22 @@
-package com.example.teumteum
+package com.example.teumteum.ui.friend
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.teumteum.databinding.FragmentFeedBinding
+import com.example.teumteum.databinding.FragmentFriendBinding
 
-class FeedFragment : Fragment() {
+class FriendFragment : Fragment() {
 
-    lateinit var binding: FragmentFeedBinding
+    lateinit var binding: FragmentFriendBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentFeedBinding.inflate(inflater, container, false)
+        binding = FragmentFriendBinding.inflate(inflater, container, false)
 
         return binding.root
     }
