+++ app/src/main/java/com/example/teumteum/ui/myhome/MyhomeFragment.kt
@@ -1,22 +1,22 @@
-package com.example.teumteum
+package com.example.teumteum.ui.myhome
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.teumteum.databinding.FragmentFriendBinding
+import com.example.teumteum.databinding.FragmentMyhomeBinding
 
-class FriendFragment : Fragment() {
+class MyhomeFragment : Fragment() {
 
-    lateinit var binding: FragmentFriendBinding
+    lateinit var binding: FragmentMyhomeBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentFriendBinding.inflate(inflater, container, false)
+        binding = FragmentMyhomeBinding.inflate(inflater, container, false)
 
         return binding.root
     }
